+++ public/index.html
@@ -8,10 +8,11 @@
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet"/>
-    <link rel="stylesheet" href="./index.css"/>
-    <script type="text/javascript" src="./index.js" defer></script>
-  </head>
+    <link rel="stylesheet" href="../index.css"/>
+    <script src="../dist/main.js" defer></script>
   
+  </head>
+
   <body>
     <main>
       <nav class="header-section">
